refactor(ledger): extract llmtrim-ledger crate (Step A0)
Move Tracker, BreakdownDb, and all row structs from llmtrim-cli into a
new crates/llmtrim-ledger (publish = false). llmtrim-cli re-exports the
types through thin wrappers so every crate::tracking:: and super::db::
call site keeps working without modification.

- crates/llmtrim-ledger/src/tracking.rs  -- verbatim from cli; removes
  the breakdown-feature gate on open_reader/open_reader_at/into_connection
  (no feature needed: the ledger crate is always fully available).
- crates/llmtrim-ledger/src/breakdown_db.rs -- verbatim from cli's
  breakdown/db.rs; crate::tracking:: refs resolve within the same crate.
- crates/llmtrim-cli/src/tracking.rs  -- replaced with pub use llmtrim_ledger::tracking::*
- crates/llmtrim-cli/src/breakdown/db.rs -- replaced with pub use llmtrim_ledger::breakdown_db::*
- All 17 tests (tracking + breakdown_db) move with the code; the cli
  suite shrinks by exactly that count, total stays 731.
- Snapshot anchor (sessions_snapshot_anchor) added in breakdown_db.rs
  before extraction and accepted; passes byte-identically in the ledger.
- cargo fmt + clippy --features intercept + nextest 731/731 green.
- cargo build -p llmtrim green.
- llmtrim-ledger coverage: breakdown_db.rs 95.33% lines, tracking.rs
  86.98% lines; uncovered paths are real-IO entrypoints (open/open_at/
  open_reader) and one-line error mappers.
