+++ crates/llmtrim-ledger/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "llmtrim-ledger"
+description = "SQLite ledger (data layer) shared by llmtrim-cli and llmtrim-tray"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+license.workspace = true
+publish = false
+
+[dependencies]
+llmtrim-core = { path = "../llmtrim-core", version = "0.4.1-dev" }
+anyhow.workspace = true
+chrono = "0.4.45"
+rusqlite = { version = "0.39", features = ["bundled"] }
+
+[dev-dependencies]
+insta = "1"
+
+[lints]
+workspace = true
